Implement i) 'see what the service can provide'
The start page isn't a true "step". It's not part of our
multi-stage "Wizard" or questionnaire which:

- has a sequence of steps around which the user can
navigate forward and back

- holds form state

- is aggregated into a "check your answers" summary

- has a final "Submitted" state

The true tasklist steps (coming next) make up the form
wizard and will be handled by the `StepsController`. They
are separate from our initial "start page" which we can
consider "Step 0".