Refactor routing for per-stage controllers
I'm thinking that the service is composed of the following
elements, some of which are "stages" within a multi-stage
questionnaire or "wizard":

- Decide whether to use service (GET start page)

- Navigate stages in questionnaire:
	- Provide dates (`dates` stage)
	- Supply registration number (`registration_number `stage)
	- Provide personal details (`personal_details` stage)

- Check answers (GET `review` controller)

- Submit application (POST to `submission` controller)

- View confirmation (GET submission )

In terms of the stages of the multi-stage questionnaire I'm
proposing to implement each with it's own controller and to give
each controller two actions:

- GET to `show` the form for that stage, including any pre-existing values
which have been provided earlier. This is a combined new/edit view.

- PUT to `update` to set or update the response(s) to a particular
stage's form

This seems like an adequate mechanism for a questionnaire where each
stage is handled on a single form view.

Controllers will be grouped in the filesystem under a `stages` namespace:

	controllers/
	   stages/
	 	 destination_stage_controller
	 	 dates_stage_controller
	 	 registration_number_stage_controller
	 	 personal_details_stage_controller

And urls can be formed with methods such as:

- `stages_destination_path`
- `stages_registration_number_stage_path`