Add a repository class to handle saving and retrieving answers
Initially we are going to use the Rails "session" to store the
answers supplied at each stage in the multi-stage wizard. This
will allow users to:

- review all their answers on a "check your answers" page
- revisit any stage and edit their answer
- submit all answers at the final submission stage

https://guides.rubyonrails.org/action_controller_overview.html#session

We wrap the interacting with the session up in a "repository" class.
This allows us to:

- design the most natural interface for our needs. We expose `save`
and `find` methods on our repository rather than expecting users to
interact directly on the `session` e.g. with `dig` and `update`

- change the storage method by making changes to the repository
implementation alone. For example if we decided to store the date
in Redis or in a database.