Strip out acceptance testing (scenario-level) comments
Now that we've written our scenarios and turned our plain
English given/when/then/and statements into method-calls and
methods, we now longer need the originally drafted spec.

However, we do retain the overall user-story as there's
value in retaining this clearly:

```
Feature: Stage 3: Supply "Spacecraft Registration Identifier"
  So that my "reg" can be checked by "the authorities"
  As a pilot
  I want to supply a validated "Spacecraft Registration Identifier"
```

When reading this spec, the purpose of the feature is
explicit.

Staying focussed on the purpose of the feature is critical
both now and in future. In the future if we note for
example that Registration IDs are no longer being checked by
the authorities, we may investigate whether this feature is
still meeting the user need as well as it once did.