Render all summary list cards from answers stored in session
In the controller we now build a CheckYourAnswers::Stage
for every stage and populate it with the user's actual
answers stored in the 'session'.

We also extend the test fixtures to verify that each
section with its questions and answers is being rendered as
expected.

This approach would need to be developed to cater for a long
questionnaire, or one in which questions and answers are
nested e.g.

```
- section
  - task
    - page
      - questions
```

But it seems like a reasonable level of abstraction for this
service.