Add default URLs for Devise
We need to set default URLs for Devise in each environment. For the local environment we can set this explicitly, but for the deployed production environment at Heroku we refer to a `HOSTNAME` environment variable to ensure we are not storing a URL that may be subject to change in multiple different places.