Collaborate with an entity presenter to build JSON representation
Our API controller is going to use a
`LandingApplicationEntity` class to build a
representation for each given `LandingApplication`.

This unit test describes that collaboration. For each
`LandingApplication`, the controller initiates an instance of
the `LandingApplicationEntity` presenter class and calls its
`represent` method.

We can use a unit test on the entity class to describe that
representation in our next commit.