Add ApiKey model to handle tokens
The `ApiKey` model will keep a persistent record in the db
of two fields:

- `api_client_name`: the name of the organisation which
  is being given access to the API, e.g. "Moon Border Force"

- `token_digest`: a hashed version of the token (this can't
  be converted back to the actual token). See [Hashing
    Passwords: One-Way Road to Security][]

The token itself is available only via the transient
attribute `unhashed_token` at the point at which the ApiKey
is generated. On subsequent retrievals of the key, the
unhashed token can not be accessed.

We'll generate the API keys manually in console sessions and
send the credentials via a secure mechanism to the API Client
contact.

This implementation was copied (almost verbatim) from Ari
Summer's article [Building a Public Facing API with Ruby on
Rails][]

[Hashing Passwords: One-Way Road to Security]:
https://auth0.com/blog/hashing-passwords-one-way-road-to-security/

[Building a Public Facing API with Ruby on Rails]:
https://therailsrunner.com/guide-to-building-public-facing-api-with-ruby-on-rails/