API controller uses ApiClientAuthenticator to verify token
The API client sends a request with the api key (token) in
a custom header: `X-API-KEY`

This is one of mechanisms [specified in OpenAPI 3.0][]

An alternative would be to use the Bearer Token format,
using the `AUTHORIZATION` header. However, the X-API-KEY
header seems clearer, and [Bearer Token][] auth is normally
used in more complex OAUTH 2.0 flows.

Note that internally Rails represents the `X-API-KEY` header
as `HTTP_X_API_KEY`, so we use that key to pull the supplied
token from `request.headers` — and also in our controller
unit test.

[specified in OpenAPI 3.0]:
https://swagger.io/docs/specification/authentication/api-keys/

[Bearer Token]:
https://datatracker.ietf.org/doc/html/rfc6750