Remove Docker caching from CI workflow
Similar to in our Rails and Nest templates, this seems to slow CI down,
so it's probably quicker not to use a cache

The branch this commit is on was originally branched off
renovate/node-20.x for the purpose of comparison. CI had been run on
that branch multiple times, but the "Run actions/cache@v3" step couldn't
find a matching cache for whatever reason. It spends almost three
minutes building that cache at the end of each test CI job run. The
first run of this branch was - expectedly - about three minutes quicker
than the reference branch (6m13s vs 3m25s)

https://github.com/dxw/rails-template/commit/1f78864444f22c4ae4372a7a584fda1e55dba88f
https://github.com/dxw/nest-template/commit/75890e11aad58773fdd0330b2796391fba5bfac7